Content change to start page title.
According to the this doc [1], “Check if simplified expenses works for your
business” to “Check if simplified expenses could save your business money”

This is becuase it appears unclear whether ‘works’ for your business means
that it will save the business money. It could be read to mean ‘the business
is able to use them’.

This commit shows the affected/changes to checksums and regression test
artefacts as a result of the content change.

I have reviewed the changes to the regression test artefacts and it shows
the difference as a result of the aforementioned change.

At this point existing unit, regression and integration tests passed, giving
assurance that the changes have had no (interfering) affect to other smart
answers.

[1] https://docs.google.com/document/d/1vyWhULg4cVGd4DJnhfBsbEtcOz077VeozWKSSdu0kSg/edit